fix the circular dependency
