+++ packages/js/src/fabric/index.ts
@@ -4,3 +4,4 @@ export * from './SignalWire'
 export * from './IncomingCallManager'
 export * from './ConversationAPI'
 export * from './types'
+export { isFabricRoomSession } from './FabricRoomSession'
